pyproject.toml
[project]
name = "ftl-quantum"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.10.8",
    "python-dotenv>=1.2.1",
    "qiskit>=2.3.0",
    "qiskit-ibm-runtime>=0.45.1",
]
